clip_services/requirements.txt
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
fastapi>=0.100.0
python-multipart>=0.0.6
uvicorn>=0.22.0
pillow>=9.5.0
numpy<2.0.0
pydantic>=2.0.0
requests>=2.31.0  # 用于健康检查 